Struct pallet_plasma::Module

pub struct Module<T: Trait>(_);

Implementations

impl<T: Trait + 'static> Module<T>

pub fn aggregator_address<K: EncodeLike<T::AccountId>>(key: K) -> T::AccountId

Single aggregator address: AggregatorId

pub fn current_block<K: EncodeLike<T::AccountId>>(key: K) -> T::BlockNumber

Current block number of commitment chain: BlockNumber

pub fn blocks<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> T::Hash where
KArg1: EncodeLike<T::AccountId>,
KArg2: EncodeLike<T::BlockNumber>,

pub fn erc20<K: EncodeLike<T::AccountId>>(key: K) -> T::AccountId

mapping from Plapps address to ERC20 based contract address.

pub fn state_update_predicate<K: EncodeLike<T::AccountId>>(
key: K
) -> T::AccountId

mapping from Plapps address to StateUpdate predicate address.

pub fn exit_predicate<K: EncodeLike<T::AccountId>>(key: K) -> T::AccountId

mapping from Plapps address to Exit predicate address.

pub fn exit_deposit_predicate<K: EncodeLike<T::AccountId>>(
key: K
) -> T::AccountId

mapping from Plapps address to ExitDeposit predicate address.

pub fn total_deposited<K: EncodeLike<T::AccountId>>(key: K) -> BalanceOf<T>

TotalDeposited is the most right coin id which has been deposited.

pub fn deposited_ranges<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> RangeOf<T> where
KArg1: EncodeLike<T::AccountId>,
KArg2: EncodeLike<BalanceOf<T>>,

pub fn checkpoints<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> bool where
KArg1: EncodeLike<T::AccountId>,
KArg2: EncodeLike<T::Hash>,

pub fn payout<K: EncodeLike<T::AccountId>>(key: K) -> T::AccountId

predicate address => payout address

impl<T: Trait> Module<T>

Public callable Plasma commitment module methods.

pub fn retrieve(
plapps_id: T::AccountId,
block_number: T::BlockNumber
) -> T::Hash

pub fn verify_inclusion(
plapps_id: T::AccountId,
leaf: T::Hash,
token_address: T::AccountId,
range: RangeOf<T>,
inclusion_proof: InclusionProofOf<T>,
block_number: T::BlockNumber
) -> DispatchResultT<bool>

verifyInclusion method verifies inclusion of message in Double Layer Tree. The message has range and token address and these also must be verified. Please see https://docs.plasma.group/projects/spec/en/latest/src/01-core/double-layer-tree.html for further details.

@param leaf a message to verify its inclusion

@param token_address token address of the message

@param range range of the message

@param inclusion_proof The proof data to verify inclusion

@param block_number block number where the Merkle root is stored

pub fn verify_inclusion_with_root(
leaf: T::Hash,
token_address: T::AccountId,
range: RangeOf<T>,
inclusion_proof: InclusionProofOf<T>,
root: T::Hash
) -> DispatchResultT<bool>

impl<T: Trait> Module<T>

Public callable Plasma deposit module methods.

pub fn bare_extend_deposited_ranges(
plapps_id: &T::AccountId,
amount: BalanceOf<T>
)

pub fn bare_remove_deposited_range(
plapps_id: &T::AccountId,
range: &RangeOf<T>,
deposited_range_id: &BalanceOf<T>
) -> DispatchResult

pub fn bare_finalize_exit(
plapps_id: &T::AccountId,
exit_property: &PropertyOf<T>,
deposited_range_id: &BalanceOf<T>
) -> DispatchResultT<StateUpdateOf<T>>

bare_finalize_exit called by this module.

@param _exitProperty A property which is instance of exit predicate and its inputs are range and StateUpdate that exiting account wants to withdraw. _exitProperty can be a property of ether ExitPredicate or ExitDepositPredicate.

@param _depositedRangeId Id of deposited range

@return return StateUpdate of exit property which is finalized.

@dev The steps of finalizeExit.

Serialize exit property

check the property is decided by Adjudication Contract.

Transfer asset to payout contract corresponding to StateObject.

Please alse see https://docs.plasma.group/projects/spec/en/latest/src/02-contracts/deposit-contract.html#finalizeexit

pub fn verify_exit_property(
plapps_id: &T::AccountId,
exit_property: &PropertyOf<T>
) -> DispatchResultT<StateUpdateOf<T>>

@dev verify StateUpdate in Exit property. _exitProperty must be instance of ether ExitPredicate or ExitDepositPredicate. if _exitProperty is instance of ExitDepositPredicate, check _exitProperty.su is subrange of _exitProperty.checkpoint.